requirements.txt
tweepy==4.14.0
pillow==10.0.0
python-dotenv 